Atualizações gerais de controle
- Os parâmetros PID agora são inteiros ao invés de floats
- Os valores PID agora são inicializados junto do robô com valores que fazem mais sentido
- Atualização nas funções de controle dos sensores do robô
- Adição da função de estimativa da posição da linha usando apenas dois sensores

+++ include/sharedTypes.h
@@ -32,10 +32,10 @@ typedef struct PIDControl {
     float kp;
     float ki;
     float kd;
-    float integral;
-    float derivative;
-    float proportional;
-    uint powerDiff;
+    int integral;
+    int derivative;
+    int proportional;
+    int powerDiff;
 } PIDControl;
 
 #endif
